requirements.txt
GitPython
matplotlib
pydriller
sip
pandas
PyQt-builder
pyqt6
PyQt6_sip
coverage
seaborn
javalang
astor
python-dateutil